Handling multi-value input that is required
Prior to this commit, if we had a required field that was also a multi-
value input, when you would click on the "Add" the empty field would
still have a `required="required"` property; This property would then
mean a modern browser would not allow you to submit your form.

By removing the "required" property of the appended input, the form can
now be submitted.

The assumption is if I have a multi-value field that is required, that
means one or more elements are required.